rename: Specific에서 Detail로 이름 변경
