+++ client/src/pages/BootCampDetail.style.ts
